refactor: view model 패키지 위치 변경
- 이전 : view model 이 feature 외부에 존재
- 이후 : view model 을 기능별 패키지 안으로 이동
