+++ app/src/main/java/woowacourse/shopping/feature/detail/ProductDetailActivity.kt
@@ -12,9 +12,10 @@ import woowacourse.shopping.data.cart.CartDummyRepository
 import woowacourse.shopping.data.product.ProductDummyRepository
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.feature.cart.CartActivity
-import woowacourse.shopping.viewmodel.CartViewModel
-import woowacourse.shopping.viewmodel.CartViewModelFactory
-import woowacourse.shopping.viewmodel.ProductViewModel
+import woowacourse.shopping.feature.cart.viewmodel.CartViewModel
+import woowacourse.shopping.feature.cart.viewmodel.CartViewModelFactory
+import woowacourse.shopping.feature.detail.viewmodel.ProductViewModel
+import woowacourse.shopping.feature.detail.viewmodel.ProductViewModelFactory
 
 class ProductDetailActivity : AppCompatActivity() {
     private val binding by lazy { ActivityProductDetailBinding.inflate(layoutInflater) }
