+++ app/src/test/java/woowacourse/shopping/feature/detail/viewmodel/ProductViewModelTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.viewmodel
+package woowacourse.shopping.feature.detail.viewmodel
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
@@ -9,6 +9,9 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.api.function.Executable
 import woowacourse.shopping.data.product.ProductDummyRepository
 import woowacourse.shopping.data.product.ProductRepository
+import woowacourse.shopping.feature.InstantTaskExecutorExtension
+import woowacourse.shopping.feature.detail.viewmodel.ProductViewModel
+import woowacourse.shopping.feature.getOrAwaitValue
 import woowacourse.shopping.imageUrl
 import woowacourse.shopping.price
 import woowacourse.shopping.title
@@ -17,12 +20,11 @@ import java.lang.IllegalArgumentException
 @ExtendWith(InstantTaskExecutorExtension::class)
 class ProductViewModelTest {
     private lateinit var viewModel: ProductViewModel
-    private lateinit var productRepository: ProductRepository
+    private val productRepository: ProductRepository = ProductDummyRepository
 
     @BeforeEach
     fun setUp() {
-        viewModel = ProductViewModel(ProductDummyRepository)
-        productRepository = ProductDummyRepository
+        viewModel = ProductViewModel(productRepository)
         productRepository.deleteAll()
     }
